+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
 
+## [0.4.0] - 2020-01-08
+- Added `gameObject` property to VGO_nodes extension.
+- Added `enabled` property to collider.
+
 ## [0.3.0] - 2020-01-01
 Initial release of the UniVGO.